web/lesson10/requirements.txt
django==4.2.5
pillow==10.0.1
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
bs4==0.0.1
lxml==4.9.3
pymongo==4.5.0
django-environ==0.11.2